+++ lib/index.ts
@@ -1,37 +1,12 @@
 import uuid from "uuid/v4";
 import { symmetricWrap } from "@botmock-api/utils";
 import { DEFAULT_INTENTS } from "../templates";
-// import { Botmock } from "@botmock-api/client";
-
-enum DelegationStrategies {
-  ALWAYS = "ALWAYS",
-  SKILL_RESPONSE = "SKILL_RESPONSE",
-}
-
-type ObjectLike<T> = { [resourceName: string]: T; };
-
-export namespace SkillsKit {
-  export type InteractionModel = {
-    languageModel: {
-      invocationName: string;
-      intents?: any[];
-      types?: { values: { name: { value: string; synonyms: string[]; }; }[]; }[];
-    };
-    dialog?: {
-      delegationStrategy?: string;
-      intents: any[];
-    };
-    prompts?: any;
-  };
-}
-
-export namespace Botmock {
-  export type Slot = {
-    samples: string[];
-  };
-  export type Intent = any;
-  export type Utterance = any;
-}
+import {
+  SkillsKit,
+  Botmock,
+  ObjectLike,
+  DelegationStrategies,
+} from "./types";
 
 /**
  * @param data public API response payload
